moved tag cache
