+++ packages/open-next/src/types/open-next.ts
@@ -4,9 +4,9 @@ import type { ReadableStream } from "node:stream/web";
 import type { StreamCreator } from "http/index.js";
 
 import type { WarmerEvent, WarmerResponse } from "../adapters/warmer-function";
-import type { TagCache } from "../cache/tag/types";
 import type { IncrementalCache } from "../overrides/incrementalCache/types";
 import type { Queue } from "../overrides/queue/types";
+import type { TagCache } from "../overrides/tagCache/types";
 
 export type BaseEventOrResult<T extends string = string> = {
   type: T;
